Added notify option that gets passed to browserSync.
Sometimes it's annoying that browserSync overlays a change notification in the browser for every change.
This can be supressed by setting the BrowserSync option 'notify' to false.

This change adds a 'notify' boolean command-line option that has a default of true. This option gets passed
to browserSync and it means that notification can be turned off by passing --no-notify on the command-line
to lite-server